+++ src/engine/mod.rs
@@ -1,3 +1,4 @@
 pub mod context;
 pub mod main_loop;
+pub mod rendering;
 pub mod state;