+++ src/main.rs
@@ -17,10 +17,11 @@ fn main() {
     // init game state
     let state = State {};
 
-    let ctx = context::Context::context()
+    let ctx = context::ContextBuilder::new()
         .title("Valewind")
         .dimensions(800, 600)
-        .build();
+        .build()
+        .unwrap(); // TODO: handle context init error more explicitly
 
     main_loop::run(ctx, &state);
 }